Worked on developing the renderer responsible for creating the ground
Added a mesh vars object that can hold customizable vertice arrays for generating new textures. this is used for prototyping the look and feel of the ground tiles which will be dynamically generating for creating more diverse terrain, the mesh vars are just used to easily edit these variables

As well expirimented with property drawers and am now comfortable creating my own object serializations and generating the subsequent UI required to easily and accurately edit the values. This can be seen on the mesh vars Scriptable object as it has a custom struct for triangle pairs that is laid out manually to make the application of the verts easier