Completed skelton for input handler and integrated it into the event System.
Allows units to be hovered over and they will change their color to indicate they are selected. once clicked the event system can handle invoking the units selection logic.

This means that the skeleton for the unit system is done and just needs to be polished up and documentation written up for creating new units and making sure the system is extensible.