First Functional Message able to sent
Able to connect a client to the host and send the first functional message. Currently a far shot from being incomplete but demonstrates the basic principles of TCP communications and the system that will be extended to encompass all TCP communication we need to do.

As well offers a patch that gives pre-compiler directives which fix issues with building when the compiler faces editor only code during a build-time compiler run. important enough to be pushed to alpha branch.