Slight file system restructering
Moved around some files to better organize the project, basically made the unity project it's own folder so different solutions could be added to the git project without messing up the design and layout of the unity project.

as such moved documentation and readme/license etc to the top level, created folder for stand alone server project which will serve as a basis for the stand alone servers for running things like match building and common server stuffs.