Started configuration to packet system instead of string messages
Started converting from a system that only sends strings to being able to send fully fledged TCP packets with custom serialization. This interface will be developed to such a way that it will be trivial for people to write data to the network interface without having to touch any of the gross TCP and multithreading required.

Currently Message demo is broken as the functions it relyed on are being removed in favour of a more universal system