Completed movement system prototype
Added movement ability to the units.

Created a system for making new actions for a unit that is extensible and able to have new actions added onto it that are automatically formatted into the UI.

Added panel hover static to reject input while we are hovering over UI elements to prevent click through issues.