Added tracking for active player and player turn and units have owner tags now
Added a tracker in the game master that keeps track of the player turn and who the active player is.

As well made them accessible via a static instance of the game master. The unit factory now checks who the active player is when making a base unit and sets the base units owner field accordingly.